Test Metrix for violations in Current, Voltage, Active and Apparent Power (#650)
* add unit tests.
* re-organize tests.
